% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\name{Boundary}
\alias{Boundary}
\title{Class Boundary (R6)}
\value{
The ojbect of class \code{Boundary}.
}
\description{
Instantiate a \code{Boundary} object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary$new()}}
\item \href{#method-waterVolume}{\code{Boundary$waterVolume()}}
\item \href{#method-clone}{\code{Boundary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$new(
  boundaryIdx,
  inputCellIdx,
  outputCellIdx,
  proportionOfInputCellFlow
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{is the index for the boundary.}

\item{\code{inputCellIdx}}{is the index for the upstream cell.}

\item{\code{outputCellIdx}}{is the index for the downstream cell.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterVolume"></a>}}
\subsection{Method \code{waterVolume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$waterVolume(upstreamCell, downstreamCell, additionalInput = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
