% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary_Reaction_Solutes.R
\name{Boundary_Reaction_Solute_Stream}
\alias{Boundary_Reaction_Solute_Stream}
\title{Class Boundary_Reaction_Solute_Stream (R6)
A model boundary that calculates solute removal from stream cells}
\value{
Fraction of solute removed or remaining in the storage zone
}
\description{
Reaction boundary for solute from stream cells
}
\section{Super classes}{
\code{\link[SoilsToStreams:Boundary]{SoilsToStreams::Boundary}} -> \code{\link[SoilsToStreams:Boundary_Reaction_Solute]{SoilsToStreams::Boundary_Reaction_Solute}} -> \code{Boundary_Reaction_Solute_Stream}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fractionRemaining}}{Fraction of solute removed from stream cell}

\item{\code{fractionRemaining}}{Fraction of solute remaining in
stream cell}

\item{\code{damkohlerNum}}{Damkohler number for the boundary}

\item{\code{damkohlerNumStorage}}{Damkohler number for the transient storage
zone (i.e. hyporheic zone)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary_Reaction_Solute_Stream$new()}}
\item \href{#method-trade}{\code{Boundary_Reaction_Solute_Stream$trade()}}
\item \href{#method-clone}{\code{Boundary_Reaction_Solute_Stream$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="SoilsToStreams" data-topic="Boundary_Reaction_Solute" data-id="calc_fracRemoval_resTimeWtdPowerLaw">}\href{../../SoilsToStreams/html/Boundary_Reaction_Solute.html#method-calc_fracRemoval_resTimeWtdPowerLaw}{\code{SoilsToStreams::Boundary_Reaction_Solute$calc_fracRemoval_resTimeWtdPowerLaw()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="SoilsToStreams" data-topic="Boundary_Reaction_Solute" data-id="processMethod_RT_PL">}\href{../../SoilsToStreams/html/Boundary_Reaction_Solute.html#method-processMethod_RT_PL}{\code{SoilsToStreams::Boundary_Reaction_Solute$processMethod_RT_PL()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="SoilsToStreams" data-topic="Boundary_Reaction_Solute" data-id="processMethod_pcnt">}\href{../../SoilsToStreams/html/Boundary_Reaction_Solute.html#method-processMethod_pcnt}{\code{SoilsToStreams::Boundary_Reaction_Solute$processMethod_pcnt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="SoilsToStreams" data-topic="Boundary_Reaction_Solute" data-id="store">}\href{../../SoilsToStreams/html/Boundary_Reaction_Solute.html#method-store}{\code{SoilsToStreams::Boundary_Reaction_Solute$store()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a solute reaction boundary in the stream
  processing domain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute_Stream$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters inherit from Class \code{\link{Boundary_Reaction_Solute}} and thus \code{\link{Boundary}}}

\item{\code{boundaryIdx}}{String indexing the boundary}

\item{\code{currency}}{String naming the currency handled by the boundary as a character e.g., \code{water, NO3}}

\item{\code{upstreamCell}}{Cell (if one exists) upstream of the boundary}

\item{\code{timeInterval}}{Model time step}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A model boundary that calculates solute removal from stream cells in the stream processing domain
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trade"></a>}}
\if{latex}{\out{\hypertarget{method-trade}{}}}
\subsection{Method \code{trade()}}{
Calculates the trades for reaction boundaries attached
  to stream cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute_Stream$trade()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Trades for reaction boundaries attached to stream cells
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute_Stream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
