% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary_Transport_Solutes.R
\name{Boundary_Transport_Solute}
\alias{Boundary_Transport_Solute}
\title{Class Boundary_Transport_Solute (R6)
Boundary to transport solutes between cells}
\description{
Transport boundary for solutes between cells
}
\section{Super class}{
\code{\link[SoilsToStreams:Boundary]{SoilsToStreams::Boundary}} -> \code{Boundary_Transport_Solute}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{load}}{Load (amount per time) of solute moved through this
boundary}

\item{\code{amount}}{Amount of solute (mass or mols) moved through this
boundary}

\item{\code{linkedBound}}{The boundary containing the water that is
advecting the solute in this boundary}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary_Transport_Solute$new()}}
\item \href{#method-trade_static}{\code{Boundary_Transport_Solute$trade_static()}}
\item \href{#method-clone}{\code{Boundary_Transport_Solute$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a transport boundary for solutes between
  cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Solute$new(..., linkedBound, load)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters inherit from Class \code{\link{Boundary}}}

\item{\code{linkedBound}}{Water boundary to which this solute boundary is
linked}

\item{\code{load}}{The rate ([mols per time] or [mass per time]) at which
solute moves through this boundary}

\item{\code{boundaryIdx}}{String indexing the boundary}

\item{\code{currency}}{String naming the currency handled by the boundary as
a character e.g., \code{water, NO3}}

\item{\code{upstreamCell}}{Cell (if one exists) upstream of the boundary}

\item{\code{downstreamCell}}{Cell (if one exists) downstream of the boundary}

\item{\code{timeInterval}}{Model time step}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A transport boundary for solutes between cells
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trade_static"></a>}}
\if{latex}{\out{\hypertarget{method-trade_static}{}}}
\subsection{Method \code{trade_static()}}{
Calculate the amount of solute to pass through the
  boundary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Solute$trade_static()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updates the \code{load} and \code{amount} in the boundary.
  Returns a list of length 2 corresponding to both \code{load} and
  \code{amount}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Solute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
