% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary_Transport_Water.R
\name{Boundary_Transport_Water_Stream}
\alias{Boundary_Transport_Water_Stream}
\title{Class Boundary_Transport_Water_Stream (R6)
A model boundary that transports water to and from stream cells}
\description{
Transport boundary for water between stream cells
}
\section{Super classes}{
\code{\link[SoilsToStreams:Boundary]{SoilsToStreams::Boundary}} -> \code{\link[SoilsToStreams:Boundary_Transport_Water]{SoilsToStreams::Boundary_Transport_Water}} -> \code{Boundary_Transport_Water_Stream}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelVelocity}}{Mean velocity of water in the channel compartment}

\item{\code{channelResidenceTime}}{Mean residence time of water in the channel compartment}

\item{\code{hydraulicLoad}}{The hydraulic load of water in the channel compartment}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary_Transport_Water_Stream$new()}}
\item \href{#method-populateDependencies}{\code{Boundary_Transport_Water_Stream$populateDependencies()}}
\item \href{#method-trade_static}{\code{Boundary_Transport_Water_Stream$trade_static()}}
\item \href{#method-clone}{\code{Boundary_Transport_Water_Stream$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a water transport boundary in the stream processing domain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{discharge}}{Rate of water discharge (a.k.a. Q)}

\item{\code{boundaryIdx}}{String indexing the boundary}

\item{\code{currency}}{String naming the currency handled by the boundary as a character e.g., \code{water, NO3}}

\item{\code{boundarySuperClass}}{String indicating the super class of the boundary, e.g., \code{transport} or \code{reaction}}

\item{\code{upstreamCell}}{Cell (if one exists) upstream of the boundary}

\item{\code{downstreamCell}}{Cell (if one exists) downstream of the boundary}

\item{\code{timeInterval}}{Model time step}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A model boundary that transports water in the stream processing domain
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populateDependencies"></a>}}
\if{latex}{\out{\hypertarget{method-populateDependencies}{}}}
\subsection{Method \code{populateDependencies()}}{
Populate boundary dependencies.  Sets the following
  \code{channelVelocity, channelResidenceTime, hydraulicLoad} based
  on the \code{discharge}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$populateDependencies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trade_static"></a>}}
\if{latex}{\out{\hypertarget{method-trade_static}{}}}
\subsection{Method \code{trade_static()}}{
Calculates the trades between the stream water cells
  using the values provided for \code{discharge}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$trade_static()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updates the \code{volume} in the cell based on
  \code{discharge}. Returns a list with two elements
  (\code{discharge, volume}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
