% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox_ConvenienceFunctions.R
\name{boundsInputFromTopoInput}
\alias{boundsInputFromTopoInput}
\title{Quickly create a list of values to instantiate new boundaries from an
  existing \code{topoInputList}}
\usage{
boundsInputFromTopoInput(topoInputList, format = "list")
}
\arguments{
\item{topoInputList}{must be a list with sublists of length three with the
following names: \code{cellIdx, inputBoundsIdx, outputBoundsIdx}.
\code{cellIdx} must be of length == 1. \code{inputBoundsIdx} and
\code{outputBoundsIdx} can be vectors of length => 1.}
}
\value{
Returns a list that is of the correct dimensions to easily
  instantiate new boundaries efficiently via \code{Boundary$new()}.
}
\description{
Quickly create a list of values to instantiate new boundaries from an
  existing \code{topoInputList}
}
