% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\name{Cell}
\alias{Cell}
\title{Class Cell (R6)}
\value{
The ojbect of class Cell
}
\description{
Instantiate a \code{Cell} object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cell$new()}}
\item \href{#method-clone}{\code{Cell$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell$new(cellIdx, inputBoundaryIdxVect, outputBoundaryIdxVect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cellIdx}}{is the index or name of the cell}

\item{\code{inputBoundaryIdxVect}}{is a vector containing the indicies of the upstream boundaries.}

\item{\code{outputBoundaryIdxVect}}{is a vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
