% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\name{Cell_Water}
\alias{Cell_Water}
\title{Class Cell_Water (R6)
Water cell}
\description{
Instantiate a water cell. Inherits from class \code{Cell}.
}
\section{Super class}{
\code{\link[SoilsToStreams:Cell]{SoilsToStreams::Cell}} -> \code{Cell_Water}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{waterVolume}}{volume of water stored in cell}

\item{\code{linkedSoluteCells}}{solute cells that are linked to the water cell}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cell_Water$new()}}
\item \href{#method-clone}{\code{Cell_Water$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new water cell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water$new(..., waterVolume = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{waterVolume}}{the volume of water in the cell}

\item{\code{cellIdx}}{Character string denoting the index for the cell}

\item{\code{processDomain}}{Character string indicating process domain of cell (soil, groundwater, or stream)}

\item{\code{currency}}{Character string with either water or name of solute}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The ojbect of class \code{Cell}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
