% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fill-n-spill_test.R
\name{Cell_Water_Soil}
\alias{Cell_Water_Soil}
\title{Class Cell_Water_Soil (R6) Soil cell}
\description{
Instantiate a soil cell. Inherits from class
  \code{\link{Cell_Water}}.
}
\section{Super classes}{
\code{\link[systERS:Cell]{systERS::Cell}} -> \code{\link[systERS:Cell_Water]{systERS::Cell_Water}} -> \code{Cell_Water_Soil}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field_capacity}}{The max volume of water that can be held within the cell.}

\item{\code{waterVolume}}{volume of water stored in cell}

\item{\code{cellVolume}}{akdj}

\item{\code{cellHeight}}{ajkdl}

\item{\code{cellWidth}}{ajd}

\item{\code{cellLength}}{akj}

\item{\code{cellPorosity}}{safh}

\item{\code{cellTortuosity}}{jafd}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cell_Water_Soil-new}{\code{Cell_Water_Soil$new()}}
\item \href{#method-Cell_Water_Soil-clone}{\code{Cell_Water_Soil$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="systERS" data-topic="Cell" data-id="populateDependencies"><a href='../../systERS/html/Cell.html#method-Cell-populateDependencies'><code>systERS::Cell$populateDependencies()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Soil-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Soil-new}{}}}
\subsection{Method \code{new()}}{
Create a new water cell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Soil$new(
  ...,
  field_capacity,
  waterVolume,
  cellLength,
  cellHeight,
  cellWidth,
  cellPorosity,
  cellTortuosity
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Inherited parameters}

\item{\code{field_capacity}}{The max volume of water that can be in the cell.}

\item{\code{waterVolume}}{Volume of water stored in cell.}

\item{\code{cellLength}}{akj}

\item{\code{cellHeight}}{ajkdl}

\item{\code{cellWidth}}{ajd}

\item{\code{cellPorosity}}{safh}

\item{\code{cellTortuosity}}{jafd}

\item{\code{cellIdx}}{Character string denoting the index for the cell}

\item{\code{processDomain}}{Character string indicating process domain of
cell (soil, groundwater, or stream)}

\item{\code{currency}}{Character string with either water or name of solute}

\item{\code{cellVolume}}{akdj}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object of class \code{Cell_Water_Soil}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Soil-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Soil-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Soil$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
