% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\name{Cell_Water_Stream}
\alias{Cell_Water_Stream}
\title{Class Cell_Water_Stream (R6)
A water cell in the stream processing domain}
\description{
Instantiate a water cell. Inherits from class \code{Cell_Water}.
}
\section{Super classes}{
\code{\link[SoilsToStreams:Cell]{SoilsToStreams::Cell}} -> \code{\link[SoilsToStreams:Cell_Water]{SoilsToStreams::Cell_Water}} -> \code{Cell_Water_Stream}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelWidth}}{Average width of stream channel in cell}

\item{\code{channelLength}}{Average length of channel in stream cell}

\item{\code{channelArea}}{Area of surface water in stream cell}

\item{\code{channelDepth}}{Average depth of channel in stream cell}

\item{\code{waterVolume}}{the volume of water in the cell calculated from
the \code{channelLength, channelArea, and channelDepth} params}

\item{\code{channelArea}}{the area of the stream channel surface of the
water in the cell}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{waterVolume}}{the volume of water in the cell calculated from
the \code{channelLength, channelArea, and channelDepth} params}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cell_Water_Stream$new()}}
\item \href{#method-clone}{\code{Cell_Water_Stream$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a \code{Class Cell_Water_Stream} object.
  Class \code{Class Cell_Water_Stream} inherits from class
  \code{Cell}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Stream$new(..., channelWidth, channelLength, channelDepth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channelWidth}}{the width of the stream channel surface (distance
from left bank to right bank)}

\item{\code{channelLength}}{is the length of the stream channel surface for
the cell (distance of cell from upstream to downstream)}

\item{\code{channelDepth}}{the height of the water surface above the
streambed}

\item{\code{cellIdx}}{Character string denoting the index for the cell}

\item{\code{processDomain}}{Character string indicating process domain of
cell (soil, groundwater, or stream)}

\item{\code{currency}}{Character string with either water or name of solute}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object of class \code{Cell_Water_Stream}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Stream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
