% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Outputter.R
\name{Outputter}
\alias{Outputter}
\title{Class Outputter (R6)
Outputs values from S2S Models}
\description{
Outputs values from S2S Models
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The S2S model object}

\item{\code{objectClassName}}{The class of S2S model object to extract
(either Cells or Bounds)}

\item{\code{attributesToReport}}{The names of objects associated with the
S2S cells or boundaries to output}

\item{\code{objectsToReport}}{The values of the S2S cells or
boundaries to output}

\item{\code{reportingInterval}}{A number indicating the reporting interval.
A value of \code{1} will report every timestep whereas a value of
\code{10} will report every 10th timestep.}

\item{\code{filePath}}{The string pointing to the place where the output
files will be written.}

\item{\code{destination}}{The string with the full file path and file name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Outputter$new()}}
\item \href{#method-report}{\code{Outputter$report()}}
\item \href{#method-getCellsOrBoundsByClass}{\code{Outputter$getCellsOrBoundsByClass()}}
\item \href{#method-getCellOrBoundAttributes}{\code{Outputter$getCellOrBoundAttributes()}}
\item \href{#method-timeSeriesGraph}{\code{Outputter$timeSeriesGraph()}}
\item \href{#method-clone}{\code{Outputter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an outputter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$new(
  model,
  objectClassName,
  attributesToReport,
  reportingInterval,
  filePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The S2S model object}

\item{\code{objectClassName}}{String with the class name of either the cells
or boundaries on which to report}

\item{\code{attributesToReport}}{Vector of strings containing the names of
the attributes to report}

\item{\code{reportingInterval}}{Numeric indicating how often model should report values}

\item{\code{filePath}}{String with the location for the output file to be stored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report"></a>}}
\if{latex}{\out{\hypertarget{method-report}{}}}
\subsection{Method \code{report()}}{
Report the output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Write output to file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCellsOrBoundsByClass"></a>}}
\if{latex}{\out{\hypertarget{method-getCellsOrBoundsByClass}{}}}
\subsection{Method \code{getCellsOrBoundsByClass()}}{
Gets cells or boundaries in the model by their class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$getCellsOrBoundsByClass(s2sObjects, objectClassName)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Selected cells or boundaries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCellOrBoundAttributes"></a>}}
\if{latex}{\out{\hypertarget{method-getCellOrBoundAttributes}{}}}
\subsection{Method \code{getCellOrBoundAttributes()}}{
Gets attributes of cells or boundaries based on the boundary name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$getCellOrBoundAttributes(s2sObjects, attributeName)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Selected attributes of either cells or boundaries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-timeSeriesGraph"></a>}}
\if{latex}{\out{\hypertarget{method-timeSeriesGraph}{}}}
\subsection{Method \code{timeSeriesGraph()}}{
Creates a of outputs through time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$timeSeriesGraph(s2sobject, attributeName, outputFilePathAndName)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
