% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\name{PcntRemovStrg}
\alias{PcntRemovStrg}
\title{PcntRemovStrg}
\description{
Calculates the fraction of a solute removed and remaining in the
  storage zone using a very simple approach of removing a certain percentage
  of the initial solute.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PcntRemovStrg$new()}}
\item \href{#method-clone}{\code{PcntRemovStrg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcntRemovStrg$new(pcntToRemove)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pcntToRemove}}{is the percent (0-100) of the solute to remove from the storage zone}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcntRemovStrg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
