% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\docType{data}
\name{ResTmWtdFracRemovStrg}
\alias{ResTmWtdFracRemovStrg}
\title{ResTmWtdFracRemovStrg}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResTmWtdFracRemovStrg
}
\description{
Calculates the fraction of a solute removed and
  remaining in the storage zone using the residence-time weighted fractional
  removal approach.
}
\keyword{datasets}
