% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSU_TradeCalcs.R
\name{SoluteTransportPerTime}
\alias{SoluteTransportPerTime}
\title{SoluteTransportPerTime}
\value{
the load to pass onto the downstream cell
}
\description{
Calculate load, i.e., the rate of solute mass moving from the
  upstream cell into the boundary (Mass/Time).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SoluteTransportPerTime$new()}}
\item \href{#method-clone}{\code{SoluteTransportPerTime$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoluteTransportPerTime$new(boundaryIdx, upstreamCell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoluteTransportPerTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
