% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\name{StreamBoundary}
\alias{StreamBoundary}
\title{Class StreamBoundary (R6)}
\value{
The object of class \code{Boundary}.
}
\description{
Instantiate a \code{StreamBoundary} object. Class
  \code{StreamBoundary} inherits from class \code{Boundary}. The following
  parameters will be calculated on the fly from the remaining parameters if
  they are not specified directly by the user: \code{channelArea,
  channelVelocity, channelResidenceTime, qStorage, hydraulicLoad}.
}
\section{Super class}{
\code{\link[SoilsToStreams:Boundary]{SoilsToStreams::Boundary}} -> \code{StreamBoundary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StreamBoundary$new()}}
\item \href{#method-clone}{\code{StreamBoundary$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="SoilsToStreams" data-topic="Boundary" data-id="waterVolume">}\href{../../SoilsToStreams/html/Boundary.html#method-waterVolume}{\code{SoilsToStreams::Boundary$waterVolume()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamBoundary$new(
  ...,
  qChannel,
  alpha,
  aquiferVolume,
  porosity,
  channelWidth,
  channelLength,
  channelArea,
  channelDepth,
  channelVelocity,
  channelResidenceTime,
  qStorage,
  hydraulicLoad,
  proportionOfUpstreamQ,
  tauMin,
  tauMax
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{qChannel}}{is the surface water discharge, Q, in user specified units.}

\item{\code{alpha}}{is the exponent of the power law used to represent the shape of
the residence time distribution of the hyporheic zone.}

\item{\code{aquiferVolume}}{is the volume of the stream aquifer/hyporheic zone.}

\item{\code{porosity}}{is the porosity of the streambed and hyporheic zone.}

\item{\code{channelWidth}}{is the width of the stream channel surface.}

\item{\code{channelLength}}{is the length of the stream channel surface for the
given reach represented by the \code{StreamBoundary} specified herein.}

\item{\code{channelArea}}{is the area of the stream channel surface for the given
\code{StreamBoundary}.}

\item{\code{channelDepth}}{is the height of the water surface above the streambed.}

\item{\code{channelVelocity}}{is the mean velocity of the water in the stream
channel.}

\item{\code{channelResidenceTime}}{is the mean residence time of the reach
represented by the \code{StreamBoundary} specified herein.}

\item{\code{qStorage}}{is the total hyporheic exchange rate, i.e., the rate at which
water is downwelling into and - since this model assumes steady state -
upwelling up from the hyporheic zone.}

\item{\code{hydraulicLoad}}{is the hydraulic load for the reach represented by the
\code{StreamBoundary} specified herein.}

\item{\code{proportionOfUpstreamQ}}{is the proportion of the water at the upstream
cell to enter into the stream boundary.  In cases where a boundary has only
one upstream cell, this value should be set to \code{1.0}. However, if flow
from an upstream cell is divergent between one or more downstream
boundaries, this parameter needs to be adjusted accordingly.}

\item{\code{tauMin}}{is the minimum residence time for the hyporheic zone.}

\item{\code{tauMax}}{is the maximum residence time for the hyporheic zone.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamBoundary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
