% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Topology.R
\name{Topology}
\alias{Topology}
\title{Class Topology (R6)}
\value{
The ojbect of class \code{Topology}.
}
\description{
Define the network topology of cells and boundaries
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Topology$new()}}
\item \href{#method-clone}{\code{Topology$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topology$new(topoList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topoList}}{At this point, \code{topoList} needs to be a list with
sublists wherein the first item is the \code{Cell} index, the second item
is the input/upstream/upnetwork \code{Boundary} index, and the third item
is the output/downstream/downnetwork \code{Boundary} index.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topology$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
