% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSU_UpdateCalcs.R
\name{UpdateBounds}
\alias{UpdateBounds}
\title{UpdateBounds}
\value{
A list of boundaries with updated values.
}
\description{
Update boundaries based on trades.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UpdateBounds$new()}}
\item \href{#method-clone}{\code{UpdateBounds$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateBounds$new(tradeDf, timeInterval)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tradeDf}}{A dataframe of trades.  This is generated by the WQModel$trade
method; currently, it is the first object returned in the list.}

\item{\code{timeInterval}}{The duration of the time step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateBounds$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
