% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSU_TradeCalcs.R
\name{WaterTransportPerTime}
\alias{WaterTransportPerTime}
\title{WaterTransportPerTime}
\value{
the discharge of water to pass through the boundary.
}
\description{
Calculate the discharge of water (Volume/Time).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WaterTransportPerTime$new()}}
\item \href{#method-clone}{\code{WaterTransportPerTime$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterTransportPerTime$new(modelEnvName, boundaryIdx, upstreamCellIdx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}

\item{\code{upstreamCellIdx}}{the name of the upstream cell as a character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterTransportPerTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
