% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_WQModel.R
\docType{data}
\name{WQModel}
\alias{WQModel}
\title{Class WQModel (R6)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WQModel
}
\arguments{
\item{boundsTable}{At this point, \code{booundsTable} needs to be a table
specifying the currency, type of boundary (reaction or transport), the
boundaryIdx (name as character), the upstreamCellIdx, the
downstreamCellIdx, and the calculateOrder.}

\item{cellsTable}{must contain the processDomain, cellIdx (as character) and
all other neccessary values to have a cell representing a process domain.}
}
\value{
The ojbect of class \code{WQModel}.
}
\description{
Define the WQ model and the network topology of cells and
  boundaries
}
\keyword{datasets}
