% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\name{Boundary}
\alias{Boundary}
\title{Boundary}
\description{
Boundary superclass
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary$new()}}
\item \href{#method-waterTransportPerTime}{\code{Boundary$waterTransportPerTime()}}
\item \href{#method-soluteTransportPerTime}{\code{Boundary$soluteTransportPerTime()}}
\item \href{#method-clone}{\code{Boundary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterTransportPerTime"></a>}}
\subsection{Method \code{waterTransportPerTime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$waterTransportPerTime(
  boundaryIdx,
  upstreamCell,
  additionalWaterInputOrLoss = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-soluteTransportPerTime"></a>}}
\subsection{Method \code{soluteTransportPerTime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$soluteTransportPerTime(
  boundaryIdx,
  upstreamCell,
  additionalSoluteInputOrLoss = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
