% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\name{Boundary}
\alias{Boundary}
\title{Class Boundary (R6)
Boundary that connects to cells.}
\description{
Instantiate a boundary
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{boundaryIdx}}{Character string naming the boundary}

\item{\code{currency}}{Character string naming the currency}

\item{\code{timeInterval}}{Model time step}

\item{\code{upstreamCell}}{The cell upstream of the boundary (\code{NA} if
the boundary is at the upstream edge of the model)}

\item{\code{downstreamCell}}{The cell downstream of the boundary (\code{NA} if
the boundary is at the downstream edge of the model)}

\item{\code{usModBound}}{Topologically, is the boundary at the upstream edge of the model? (TRUE/FALSE)}

\item{\code{dsModBound}}{Topologically, is the boundary at the downstream edge of the model? (TRUE/FALSE)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Boundary-new}{\code{Boundary$new()}}
\item \href{#method-Boundary-clone}{\code{Boundary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary-new"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a boundary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$new(boundaryIdx, currency, timeInterval, upstreamCell, downstreamCell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{String indexing the boundary}

\item{\code{currency}}{String naming the currency handled by the boundary as a character e.g., \code{water, NO3}}

\item{\code{timeInterval}}{Model time step}

\item{\code{upstreamCell}}{Cell (if one exists) upstream of the boundary}

\item{\code{downstreamCell}}{Cell (if one exists) downstream of the boundary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A model boundary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
