% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\name{Boundary}
\alias{Boundary}
\title{Boundary}
\description{
Boundary superclass
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary$new()}}
\item \href{#method-waterTransportPerTime}{\code{Boundary$waterTransportPerTime()}}
\item \href{#method-soluteTransportPerTime}{\code{Boundary$soluteTransportPerTime()}}
\item \href{#method-fractionalSoluteDynams}{\code{Boundary$fractionalSoluteDynams()}}
\item \href{#method-clone}{\code{Boundary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$new(
  boundaryIdx,
  currency,
  boundarySuperClass,
  upstreamCellIdx,
  downstreamCellIdx,
  calculateOrder
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}

\item{\code{currency}}{the name of the currency handled by the boundary as a character e.g., \code{H20}}

\item{\code{boundarySuperClass}}{the super class of the boundary, e.g., \code{transport} or \code{reaction}}

\item{\code{upstreamCellIdx}}{the name of the upstream cell as a character.}

\item{\code{downstreamCellIdx}}{the name of the downstream cell as a character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterTransportPerTime"></a>}}
\if{latex}{\out{\hypertarget{method-waterTransportPerTime}{}}}
\subsection{Method \code{waterTransportPerTime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$waterTransportPerTime(boundaryIdx, upstreamCellIdx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}

\item{\code{upstreamCellIdx}}{the name of the upstream cell as a character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-soluteTransportPerTime"></a>}}
\if{latex}{\out{\hypertarget{method-soluteTransportPerTime}{}}}
\subsection{Method \code{soluteTransportPerTime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$soluteTransportPerTime(boundaryIdx, upstreamCell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fractionalSoluteDynams"></a>}}
\if{latex}{\out{\hypertarget{method-fractionalSoluteDynams}{}}}
\subsection{Method \code{fractionalSoluteDynams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$fractionalSoluteDynams(
  boundaryIdx,
  upstreamCellIdx,
  removalMethod,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}

\item{\code{upstreamCellIdx}}{the name of the upstream cell as a character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
