% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\docType{data}
\name{Boundary}
\alias{Boundary}
\title{Boundary}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Boundary
}
\arguments{
\item{boundaryIdx}{the name of the boundary.}

\item{upstreamCellIdx}{the name of the upstream cell as a character.}

\item{downstreamCellIdx}{the name of the downstream cell as a character.}
}
\description{
Boundary superclass
}
\keyword{datasets}
