% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary_Reaction_Solutes.R
\name{Boundary_Reaction_Solute}
\alias{Boundary_Reaction_Solute}
\title{Class Boundary_Reaction_Solute (R6)
Reaction boundary for a solute}
\value{
Fraction of solute removed or remaining in the storage zone
}
\description{
Reaction boundary for a solute
}
\section{Super class}{
\code{\link[SoilsToStreams:Boundary]{SoilsToStreams::Boundary}} -> \code{Boundary_Reaction_Solute}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fractionRemovedStorage}}{Fraction of solute removed from storage
by the boundary}

\item{\code{fractionRemainingStorage}}{Fraction of solute remaining in
storage by the boundary}

\item{\code{mustBeOne}}{Used for an error check to ensure that the fraction
of solute removed from and remaining in a cell sum to 1}

\item{\code{startingAmount}}{Amount of solute prior to reaction}

\item{\code{amountToRemove}}{Amount of solute the reaction boundary removes}

\item{\code{amountToRemain}}{Amount of solute remaining after the reaction}

\item{\code{processDomain}}{Process domain of cell on which the reaction
boundary is interacting}

\item{\code{processMethod}}{How to process the solute, either \code{pcnt} or \code{RT-PL}}

\item{\code{tauMin}}{Minimum residence time to consider}

\item{\code{tauMax}}{Maximum residence time to consider}

\item{\code{alpha}}{Power law exponent describing the shape of the curve,
typically between -1.2 and -1.9}

\item{\code{k}}{Uptake constant for solute in units of T-1}

\item{\code{qStorage}}{Volumeteric rate of water entering the storage zone}

\item{\code{rxnVals}}{A data frame storing the key reaction boundary inputs
and outputs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary_Reaction_Solute$new()}}
\item \href{#method-calc_removal_pcnt}{\code{Boundary_Reaction_Solute$calc_removal_pcnt()}}
\item \href{#method-calc_fracRemoval_resTimeWtdPowerLaw}{\code{Boundary_Reaction_Solute$calc_fracRemoval_resTimeWtdPowerLaw()}}
\item \href{#method-calc_trade_stream}{\code{Boundary_Reaction_Solute$calc_trade_stream()}}
\item \href{#method-clone}{\code{Boundary_Reaction_Solute$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a reaction boundary for a solute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{String indexing the boundary}

\item{\code{currency}}{String naming the currency handled by the boundary as
a character e.g., \code{water, NO3}}

\item{\code{boundarySuperClass}}{String indicating the super class of the
boundary, e.g., \code{transport} or \code{reaction}}

\item{\code{upstreamCell}}{Cell (if one exists) upstream of the boundary}

\item{\code{downstreamCell}}{Cell (if one exists) downstream of the boundary}

\item{\code{timeInterval}}{Model time step}

\item{\code{processMethod}}{How to process the solute, either \code{pcnt} or \code{RT-PL}}

\item{\code{tauMin}}{Minimum residence time to consider}

\item{\code{tauMax}}{Maximum residence time to consider}

\item{\code{alpha}}{Power law exponent describing the shape of the curve,
typically between -1.2 and -1.9}

\item{\code{k}}{Uptake constant for solute in units of T-1}

\item{\code{qStorage}}{Volumeteric rate of water entering the storage zone}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The ojbect of class \code{Boundary_Reaction_Solute}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_removal_pcnt"></a>}}
\if{latex}{\out{\hypertarget{method-calc_removal_pcnt}{}}}
\subsection{Method \code{calc_removal_pcnt()}}{
Calculates the fraction solute to remove from storage
  based on either a user specified percent or a power law residence
  time weighted approach
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$calc_removal_pcnt()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pcntToRemove}}{Fraction of solute to remove from storage}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_fracRemoval_resTimeWtdPowerLaw"></a>}}
\if{latex}{\out{\hypertarget{method-calc_fracRemoval_resTimeWtdPowerLaw}{}}}
\subsection{Method \code{calc_fracRemoval_resTimeWtdPowerLaw()}}{
Calculate fraction of solute removed/remainig from a
  fully saturated storage zone using a residence-time weighted
  removal function assuming a power law residence time.  If remaining
  = TRUE, then the function calculates the fraction of the solute
  REMAINING; however if remaining = FALSE, then the function
  calculates the fraction of solute REMOVED
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$calc_fracRemoval_resTimeWtdPowerLaw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_trade_stream"></a>}}
\if{latex}{\out{\hypertarget{method-calc_trade_stream}{}}}
\subsection{Method \code{calc_trade_stream()}}{
Calculates the trades for reaction boundaries attached
  to stream cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$calc_trade_stream()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Trades for reaction boundaries attached to stream cells
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
