% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary_Reaction_Solutes.R
\name{Boundary_Reaction_Solute}
\alias{Boundary_Reaction_Solute}
\title{Class Boundary_Reaction_Solute (R6)
Reaction boundary for a solute}
\value{
Fraction of solute removed or remaining in the storage zone
}
\description{
Reaction boundary for a solute
}
\section{Super class}{
\code{\link[systERS:Boundary]{systERS::Boundary}} -> \code{Boundary_Reaction_Solute}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fractionRemovedStorage}}{Fraction of solute removed from storage
by the boundary}

\item{\code{fractionRemainingStorage}}{Fraction of solute remaining in
storage by the boundary}

\item{\code{mustBeOne}}{Used for an error check to ensure that the fraction
of solute removed from and remaining in a cell sum to 1}

\item{\code{startingAmount}}{Amount of solute prior to reaction}

\item{\code{amountToRemove}}{Amount of solute the reaction boundary removes}

\item{\code{amountToRemain}}{Amount of solute remaining after the reaction}

\item{\code{processDomain}}{Process domain of cell on which the reaction
boundary is interacting}

\item{\code{processMethodName}}{How to process the solute, either
\code{Boundary_Reaction_Solute$processMethod_RT_PL} or
\code{Boundary_Reaction_Solute$processMethod_pcnt}}

\item{\code{processMethod}}{Call to the Method  \code{pcnt} or \code{RT-PL}}

\item{\code{tauMin}}{Smallest residence time when solute processing is
possible; must be >= tauMin}

\item{\code{tauMax}}{Largest residence time when solute processing is
possible; must be <= tauMax}

\item{\code{alpha}}{Power law exponent describing the shape of the curve,
typically between -1.2 and -1.9}

\item{\code{k}}{Uptake constant for solute in units of T-1}

\item{\code{tauRxn}}{The minimum residence time when reaction can occur,
i.e., a value that specifies how long solute must be in the
reactive storage zone before it has the capacity to react.}

\item{\code{qStorage}}{Volumetric rate of water entering the storage zone}

\item{\code{volWaterInStorage}}{Volume of water in the transient storage
zone (equal to the volume of the alluvial aquifer X porosity)}

\item{\code{rxnVals}}{A data frame storing the key reaction boundary inputs
and outputs}

\item{\code{pcntToRemove}}{If process method is set to \code{pcnt}, then the
percent of solute amount (mass or mols) to remove from storage must
be specified}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Boundary_Reaction_Solute-new}{\code{Boundary_Reaction_Solute$new()}}
\item \href{#method-Boundary_Reaction_Solute-processMethod_RT_PL}{\code{Boundary_Reaction_Solute$processMethod_RT_PL()}}
\item \href{#method-Boundary_Reaction_Solute-processMethod_pcnt}{\code{Boundary_Reaction_Solute$processMethod_pcnt()}}
\item \href{#method-Boundary_Reaction_Solute-calc_fracRemoval_resTimeWtdPowerLaw}{\code{Boundary_Reaction_Solute$calc_fracRemoval_resTimeWtdPowerLaw()}}
\item \href{#method-Boundary_Reaction_Solute-store}{\code{Boundary_Reaction_Solute$store()}}
\item \href{#method-Boundary_Reaction_Solute-clone}{\code{Boundary_Reaction_Solute$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary_Reaction_Solute-new"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary_Reaction_Solute-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a reaction boundary for a solute #'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$new(
  ...,
  processMethodName,
  tauMin,
  tauMax,
  alpha,
  k,
  tauRxn,
  qStorage,
  pcntToRemove,
  volWaterInStorage
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters inherited from Class \code{\link{Boundary}}}

\item{\code{processMethodName}}{How to process the solute, either
\code{pcnt} or \code{RT-PL}}

\item{\code{tauMin}}{Minimum residence time of water to consider; bound on
lower end of power law describing the residence time distribution}

\item{\code{tauMax}}{Maximum residence time of water to consider; bound on
upper end of power law describing the residence time distribution}

\item{\code{alpha}}{Power law exponent describing the shape of the curve,
typically between -1.2 and -1.9}

\item{\code{k}}{Uptake constant for solute in units of T-1}

\item{\code{tauRxn}}{A lag value that specifies how long solute must be in
the reactive storage zone before it has the capacity to react,
i.e., this is the minimum residence time when reaction can occur.
This parameter is only passed to the \code{RT-PL} Method and cannot
be applied to the \code{pcnt} Method.  The default value is
\code{0}.}

\item{\code{qStorage}}{Volumetric rate of water entering the storage zone}

\item{\code{pcntToRemove}}{Percent of solute amount (mass or mols) to remove
from storage}

\item{\code{volWaterInStorage}}{The volume of water in the reactive storage
zone}

\item{\code{boundaryIdx}}{String indexing the boundary}

\item{\code{currency}}{String naming the currency handled by the boundary as
a character e.g., \code{water, NO3}}

\item{\code{upstreamCell}}{Cell to which the reaction boundary is connected}

\item{\code{timeInterval}}{Model time step}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object of class \code{Boundary_Reaction_Solute}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary_Reaction_Solute-processMethod_RT_PL"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary_Reaction_Solute-processMethod_RT_PL}{}}}
\subsection{Method \code{processMethod_RT_PL()}}{
Calculates the fraction solute to remove from storage
  based on a power law residence time weighted approach
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$processMethod_RT_PL()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{remaining}}{If \code{TRUE}, returns solute remaining.  If
\code{FALSE}, returns solute removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value of fraction removed and remaining in storage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary_Reaction_Solute-processMethod_pcnt"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary_Reaction_Solute-processMethod_pcnt}{}}}
\subsection{Method \code{processMethod_pcnt()}}{
Calculates the fraction solute to remove from storage
  based on  a user specified percent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$processMethod_pcnt()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Value of fraction removed and remaining in storage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary_Reaction_Solute-calc_fracRemoval_resTimeWtdPowerLaw"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary_Reaction_Solute-calc_fracRemoval_resTimeWtdPowerLaw}{}}}
\subsection{Method \code{calc_fracRemoval_resTimeWtdPowerLaw()}}{
Calculate fraction of solute removed/remainig from a
  fully saturated storage zone using a residence-time weighted
  removal function assuming a power law residence time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$calc_fracRemoval_resTimeWtdPowerLaw(
  remaining = remaining
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{remaining}}{If \code{TRUE}, returns fraction of solute remaining.  If
\code{FALSE}, returns fraction of solute removed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary_Reaction_Solute-store"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary_Reaction_Solute-store}{}}}
\subsection{Method \code{store()}}{
Runs the store method on solute cells in the model for
  reactions that remove solute from cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$store()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updated store values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boundary_Reaction_Solute-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Boundary_Reaction_Solute-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Reaction_Solute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
