% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary_Transport_Water.R
\name{Boundary_Transport_Water_Stream}
\alias{Boundary_Transport_Water_Stream}
\title{Class Boundary_Transport_Water_Stream (R6)
A model boundary that transports water to and from stream cells}
\description{
Transport boundary for water between stream cells
}
\section{Super classes}{
\code{\link[SoilsToStreams:Boundary]{SoilsToStreams::Boundary}} -> \code{\link[SoilsToStreams:Boundary_Transport_Water]{SoilsToStreams::Boundary_Transport_Water}} -> \code{Boundary_Transport_Water_Stream}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelVelocity}}{Mean velocity of water in the channel compartment}

\item{\code{channelResidenceTime}}{Mean residence time of water in the channel compartment}

\item{\code{hydraulicLoad}}{The hydraulic load of water in the channel compartment}

\item{\code{populateDependencies}}{Updates channel velocity, residence time, and hydraulic load.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boundary_Transport_Water_Stream$new()}}
\item \href{#method-populateDependenciesInternalBound}{\code{Boundary_Transport_Water_Stream$populateDependenciesInternalBound()}}
\item \href{#method-populateDependenciesExternalBound}{\code{Boundary_Transport_Water_Stream$populateDependenciesExternalBound()}}
\item \href{#method-trade}{\code{Boundary_Transport_Water_Stream$trade()}}
\item \href{#method-clone}{\code{Boundary_Transport_Water_Stream$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="SoilsToStreams" data-topic="Boundary_Transport_Water" data-id="store">}\href{../../SoilsToStreams/html/Boundary_Transport_Water.html#method-store}{\code{SoilsToStreams::Boundary_Transport_Water$store()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a water transport boundary in the stream processing domain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters inherit from Class \code{\link{Boundary_Transport_Water}} and thus \code{\link{Boundary}}}

\item{\code{discharge}}{Rate of water discharge (a.k.a. Q)}

\item{\code{boundaryIdx}}{String indexing the boundary}

\item{\code{currency}}{String naming the currency handled by the boundary as a character e.g., \code{water, NO3}}

\item{\code{upstreamCell}}{Cell (if one exists) upstream of the boundary}

\item{\code{downstreamCell}}{Cell (if one exists) downstream of the boundary}

\item{\code{timeInterval}}{Model time step}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A model boundary that transports water in the stream processing domain
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populateDependenciesInternalBound"></a>}}
\if{latex}{\out{\hypertarget{method-populateDependenciesInternalBound}{}}}
\subsection{Method \code{populateDependenciesInternalBound()}}{
Populate boundary dependencies for boundaries with
  exactly one upstream and one downstream cell.  Sets the following
  \code{channelVelocity, channelResidenceTime, hydraulicLoad} based
  on the \code{discharge}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$populateDependenciesInternalBound()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Populates boundary dependencies
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populateDependenciesExternalBound"></a>}}
\if{latex}{\out{\hypertarget{method-populateDependenciesExternalBound}{}}}
\subsection{Method \code{populateDependenciesExternalBound()}}{
Populate boundary dependencies for boundaries at the
  edge of the topology (i.e., with either one upstream or one
  downstream cell).  Sets the following \code{channelVelocity,
  channelResidenceTime, hydraulicLoad} based on the \code{discharge}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$populateDependenciesExternalBound()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Populates boundary dependencies
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trade"></a>}}
\if{latex}{\out{\hypertarget{method-trade}{}}}
\subsection{Method \code{trade()}}{
Calculates the trades between the stream water cells
  using the values provided for \code{discharge}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$trade()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updates the \code{volume} in the cell based on
  \code{discharge}. Returns a list with two elements
  (\code{discharge, volume}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boundary_Transport_Water_Stream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
