% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSU_TradeCalcs.R
\name{CalcFractionalSoluteDynams}
\alias{CalcFractionalSoluteDynams}
\title{CalcFractionalSoluteDynams}
\value{
the fraction of solute removed and remaining in the cell by the
  boundary.
}
\description{
Calculate the fraction of solute removed and remaining.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CalcFractionalSoluteDynams$new()}}
\item \href{#method-clone}{\code{CalcFractionalSoluteDynams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcFractionalSoluteDynams$new(
  boundaryIdx,
  upstreamCellIdx,
  removalMethod,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundaryIdx}}{the name of the boundary.}

\item{\code{removalMethod}}{the method by which solute is removed from the cell.
Currently, only \code{RT-PL} and \code{pcnt} is supported.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcFractionalSoluteDynams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
