% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSU_StoreCalcs.R
\name{CalcStores}
\alias{CalcStores}
\title{CalcStores}
\value{
A list containing 2 tables--one corresponding to H2O and the other
  NO3--with mass balances for each cell for each currency.
}
\description{
Calculate the mass balances for water and solute.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CalcStores$new()}}
\item \href{#method-doCalc}{\code{CalcStores$doCalc()}}
\item \href{#method-clone}{\code{CalcStores$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcStores$new(cells, bounds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{A list of cells in the model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doCalc"></a>}}
\subsection{Method \code{doCalc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcStores$doCalc(tradeTable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tradeTable}}{A table with the neccessary trade information to calculate
the stores.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcStores$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
