% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\name{Cell}
\alias{Cell}
\title{Class Cell (R6)
Model cell}
\description{
Instantiate a \code{Cell}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cellIdx}}{Cell index}

\item{\code{processDomain}}{Process domain of the cell}

\item{\code{currency}}{Currency of the cell}

\item{\code{linkedBoundsList}}{List of boundaries linked to cell}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cell-new}{\code{Cell$new()}}
\item \href{#method-Cell-populateDependencies}{\code{Cell$populateDependencies()}}
\item \href{#method-Cell-clone}{\code{Cell$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cell-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a \code{Cell} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell$new(cellIdx, processDomain, currency)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cellIdx}}{Character string denoting the index for the cell}

\item{\code{processDomain}}{Character string indicating process domain of cell (soil, groundwater, or stream)}

\item{\code{currency}}{Character string with either water or name of solute}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The ojbect of class \code{Cell}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell-populateDependencies"></a>}}
\if{latex}{\out{\hypertarget{method-Cell-populateDependencies}{}}}
\subsection{Method \code{populateDependencies()}}{
Placeholder for populate dependencies methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell$populateDependencies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cell-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
