% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\docType{data}
\name{Cell}
\alias{Cell}
\title{Class Cell (R6)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cell
}
\arguments{
\item{cellIdx}{is the index for the cell.}
}
\value{
The ojbect of class \code{Cell}.
}
\description{
Instantiate a \code{Cell} object.
}
\keyword{datasets}
