% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fill-n-spill_test.R
\name{Cell_Water_Soil}
\alias{Cell_Water_Soil}
\title{Class Cell_Water_Soil (R6) Soil cell}
\description{
Instantiate a soil cell. Inherits from class
  \code{\link{Cell_Water}}.
}
\section{Super classes}{
\code{\link[systERS:Cell]{systERS::Cell}} -> \code{\link[systERS:Cell_Water]{systERS::Cell_Water}} -> \code{Cell_Water_Soil}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fieldCapacity}}{The max volume of water that can be held within the cell.}

\item{\code{cellVolume}}{The volume of the soil cell calculate from from
the \code{channelLength, channelWidth, channelHeight} parameters.}

\item{\code{cellHeight}}{The height of the soil cell.}

\item{\code{cellWidth}}{The width of the soil cell.}

\item{\code{cellLength}}{The length of the soil cell.}

\item{\code{cellPorosity}}{The porosity of the soil cell.}

\item{\code{cellMatricPotential}}{The matric potential of the soil cell.}

\item{\code{cellGravimetricPotential}}{The total pressure gradient for the soil cell.}

\item{\code{cellSoilType}}{The soil type of the cell. For notation purposes only.}

\item{\code{waterVolume}}{The volume of water stored in the soil cell.}

\item{\code{cellInput}}{The volume of water entering the soil cell.}

\item{\code{cellSpillOver}}{The volume of water exiting the soil cell.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cell_Water_Soil-new}{\code{Cell_Water_Soil$new()}}
\item \href{#method-Cell_Water_Soil-populateDependencies}{\code{Cell_Water_Soil$populateDependencies()}}
\item \href{#method-Cell_Water_Soil-update}{\code{Cell_Water_Soil$update()}}
\item \href{#method-Cell_Water_Soil-clone}{\code{Cell_Water_Soil$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Soil-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Soil-new}{}}}
\subsection{Method \code{new()}}{
Create a new water cell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Soil$new(
  ...,
  cellLength,
  cellHeight,
  cellWidth,
  cellPorosity,
  cellSoilType,
  initWaterVolume
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Inherited parameters}

\item{\code{cellLength}}{The length of the soil cell.}

\item{\code{cellHeight}}{The height of the soil cell.}

\item{\code{cellWidth}}{The width of the soil cell.}

\item{\code{cellPorosity}}{The porosity of the soil cell.}

\item{\code{cellSoilType}}{The soil type of the cell. For notation purposes only.}

\item{\code{fieldCapacity}}{The max volume of water that can be in the cell.}

\item{\code{cellIdx}}{Character string denoting the index for the cell}

\item{\code{processDomain}}{Character string indicating process domain of
cell (soil, groundwater, or stream)}

\item{\code{currency}}{Character string with either water or name of solute}

\item{\code{cellVolume}}{The volume of the soil cell.}

\item{\code{cellMatricPotential}}{The matric potential of the soil cell.}

\item{\code{cellGravimetricPotential}}{The total pressure gradient for the soil cell.}

\item{\code{waterVolume}}{The volume of water already present in the cell.}

\item{\code{cellInput}}{The volume of water entering the soil cell.}

\item{\code{cellSpillOver}}{The volume of water exiting the soil cell.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object of class \code{Cell_Water_Soil}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Soil-populateDependencies"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Soil-populateDependencies}{}}}
\subsection{Method \code{populateDependencies()}}{
Populates the fields in the cells that depend on
  boundaries being instantiated before the trade, store, update
  sequence can be run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Soil$populateDependencies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updates cell values for \code{cellSpillOver}
based on cell values (\code{fieldCapacity, waterVolume}) and
  upstream/downstream boundary values (\code{cellInput}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Soil-update"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Soil-update}{}}}
\subsection{Method \code{update()}}{
Runs the update method on all cells of class
  \code{Cell_Water_Soil}.  In this current version of the model,
  this simply adjusts the height of the water in the stream cell
  based on the water volume, i.e., it holds the channel area constant
  with changes in discharge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Soil$update()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updates cell values based on trades and stores.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Soil-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Soil-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Soil$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
