% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\name{Cell_Water_Stream}
\alias{Cell_Water_Stream}
\title{Class Cell_Water_Stream (R6)
A water cell in the stream processing domain}
\description{
Instantiate a water cell. Inherits from class \code{Cell_Water}.
}
\section{Super classes}{
\code{\link[systERS:Cell]{systERS::Cell}} -> \code{\link[systERS:Cell_Water]{systERS::Cell_Water}} -> \code{Cell_Water_Stream}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelWidth}}{Average width of stream channel in cell}

\item{\code{channelLength}}{Average length of channel in stream cell}

\item{\code{channelArea}}{Area of surface water in stream cell}

\item{\code{channelDepth}}{Average depth of channel in stream cell}

\item{\code{waterVolume}}{the volume of water in the cell calculated from
the \code{channelLength, channelArea, and channelDepth} params}

\item{\code{channelResidenceTime}}{Mean residence time of water in the channel compartment}

\item{\code{hydraulicLoad}}{The hydraulic load of water in the channel compartment}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cell_Water_Stream-new}{\code{Cell_Water_Stream$new()}}
\item \href{#method-Cell_Water_Stream-populateDependencies}{\code{Cell_Water_Stream$populateDependencies()}}
\item \href{#method-Cell_Water_Stream-update}{\code{Cell_Water_Stream$update()}}
\item \href{#method-Cell_Water_Stream-clone}{\code{Cell_Water_Stream$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Stream-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Stream-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a \code{Class Cell_Water_Stream} object.
  Class \code{Class Cell_Water_Stream} inherits from class
  \code{Cell}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Stream$new(..., channelWidth, channelLength, channelDepth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters inherit from Class \code{\link{Cell_Water}} and
thus \code{\link{Cell}}}

\item{\code{channelWidth}}{the width of the stream channel surface (distance
from left bank to right bank)}

\item{\code{channelLength}}{is the length of the stream channel surface for
the cell (distance of cell from upstream to downstream)}

\item{\code{channelDepth}}{the height of the water surface above the
streambed}

\item{\code{cellIdx}}{Character string denoting the index for the cell}

\item{\code{processDomain}}{Character string indicating process domain of
cell (soil, groundwater, or stream)}

\item{\code{currency}}{Character string with either water or name of solute}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object of class \code{Cell_Water_Stream}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Stream-populateDependencies"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Stream-populateDependencies}{}}}
\subsection{Method \code{populateDependencies()}}{
Populates the fields in the cells that depend on
  boundaries being instantiated before the trade, store, update
  sequence can be run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Stream$populateDependencies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updates cell values for \code{channelResidenceTime,
  hydraulicLoad} based on cell values (\code{channelLength}) and
  upstream/downstream boundary values (\code{channelVelocity}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Stream-update"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Stream-update}{}}}
\subsection{Method \code{update()}}{
Runs the update method on all cells of class
  \code{Cell_Water_Stream}.  In this current version of the model,
  this simply adjusts the height of the water in the stream cell
  based on the water volume, i.e., it holds the channel area constant
  with changes in discharge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Stream$update()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updates cell values based on trades and stores.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cell_Water_Stream-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cell_Water_Stream-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cell_Water_Stream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
