% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Outputter.R
\name{Outputter}
\alias{Outputter}
\title{Class Outputter (R6)
Outputs values from systERS Models}
\description{
Outputs values from systERS Models
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The systERS model object}

\item{\code{objectClassName}}{The class of systERS model object to extract
(either Cells or Bounds)}

\item{\code{attributesToReport}}{The names of objects associated with the
systERS cells or boundaries to output}

\item{\code{objectsToReport}}{The values of the systERS cells or
boundaries to output}

\item{\code{reportingInterval}}{A number indicating the reporting interval.
A value of \code{1} will report every timestep whereas a value of
\code{10} will report every 10th timestep.}

\item{\code{filePath}}{The string pointing to the place where the output
files will be written.}

\item{\code{destination}}{The string with the full file path and file name.
File type and extension must be compatible with
\code{write.table()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Outputter-new}{\code{Outputter$new()}}
\item \href{#method-Outputter-report}{\code{Outputter$report()}}
\item \href{#method-Outputter-getCellsOrBoundsByClass}{\code{Outputter$getCellsOrBoundsByClass()}}
\item \href{#method-Outputter-getCellOrBoundAttributes}{\code{Outputter$getCellOrBoundAttributes()}}
\item \href{#method-Outputter-timeSeriesGraph}{\code{Outputter$timeSeriesGraph()}}
\item \href{#method-Outputter-clone}{\code{Outputter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outputter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Outputter-new}{}}}
\subsection{Method \code{new()}}{
Create an outputter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$new(
  model,
  objectClassName,
  attributesToReport,
  reportingInterval,
  filePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The systERS model object}

\item{\code{objectClassName}}{String with the class name of either the cells
or boundaries on which to report}

\item{\code{attributesToReport}}{Vector of strings containing the names of
the attributes to report}

\item{\code{reportingInterval}}{Numeric indicating how often model should report values}

\item{\code{filePath}}{String with the location for the output file to be stored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outputter-report"></a>}}
\if{latex}{\out{\hypertarget{method-Outputter-report}{}}}
\subsection{Method \code{report()}}{
Report the output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Write output to file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outputter-getCellsOrBoundsByClass"></a>}}
\if{latex}{\out{\hypertarget{method-Outputter-getCellsOrBoundsByClass}{}}}
\subsection{Method \code{getCellsOrBoundsByClass()}}{
Gets cells or boundaries in the model by their class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$getCellsOrBoundsByClass(systERSObjects, objectClassName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{systERSObjects}}{Objects of class systERS}

\item{\code{objectClassName}}{String with the class name of either the cells
or boundaries on which to report}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Selected cells or boundaries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outputter-getCellOrBoundAttributes"></a>}}
\if{latex}{\out{\hypertarget{method-Outputter-getCellOrBoundAttributes}{}}}
\subsection{Method \code{getCellOrBoundAttributes()}}{
Gets attributes of cells or boundaries based on the
  boundary name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$getCellOrBoundAttributes(systERSObjects, attributeName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{systERSObjects}}{Objects of class systERS}

\item{\code{attributeName}}{Name of attribute}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Selected attributes of either cells or boundaries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outputter-timeSeriesGraph"></a>}}
\if{latex}{\out{\hypertarget{method-Outputter-timeSeriesGraph}{}}}
\subsection{Method \code{timeSeriesGraph()}}{
Creates a of outputs through time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$timeSeriesGraph(systERSobject, attributeName, outputFilePathAndName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{systERSobject}}{An object of class systERS}

\item{\code{attributeName}}{Name of attribute}

\item{\code{outputFilePathAndName}}{File path where outputter will write
results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outputter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Outputter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outputter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
