% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\docType{data}
\name{PcntRemovStrg}
\alias{PcntRemovStrg}
\title{PcntRemovStrg}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PcntRemovStrg
}
\arguments{
\item{pcntToRemove}{is the percent (0-100) of the solute to remove from the storage zone}
}
\description{
Calculates the fraction of a solute removed and remaining in the
  storage zone using a very simple approach of removing a certain percentage
  of the initial solute.
}
\keyword{datasets}
