% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Boundary.R
\name{ResTmWtdFracRemovStrg}
\alias{ResTmWtdFracRemovStrg}
\title{ResTmWtdFracRemovStrg}
\description{
Calculates the fraction of a solute removed and
  remaining in the storage zone using the residence-time weighted fractional
  removal approach.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResTmWtdFracRemovStrg$new()}}
\item \href{#method-frequency}{\code{ResTmWtdFracRemovStrg$frequency()}}
\item \href{#method-density}{\code{ResTmWtdFracRemovStrg$density()}}
\item \href{#method-fractionRemoved}{\code{ResTmWtdFracRemovStrg$fractionRemoved()}}
\item \href{#method-integrateFractionRemoved}{\code{ResTmWtdFracRemovStrg$integrateFractionRemoved()}}
\item \href{#method-fractionRemaining}{\code{ResTmWtdFracRemovStrg$fractionRemaining()}}
\item \href{#method-integrateFractionRemaining}{\code{ResTmWtdFracRemovStrg$integrateFractionRemaining()}}
\item \href{#method-clone}{\code{ResTmWtdFracRemovStrg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$new(alpha, k, tauMin, tauMax, freqMin = tauMax^-alpha)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frequency"></a>}}
\if{latex}{\out{\hypertarget{method-frequency}{}}}
\subsection{Method \code{frequency()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$frequency(tau)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-density"></a>}}
\if{latex}{\out{\hypertarget{method-density}{}}}
\subsection{Method \code{density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$density(tau)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fractionRemoved"></a>}}
\if{latex}{\out{\hypertarget{method-fractionRemoved}{}}}
\subsection{Method \code{fractionRemoved()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$fractionRemoved(tau)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-integrateFractionRemoved"></a>}}
\if{latex}{\out{\hypertarget{method-integrateFractionRemoved}{}}}
\subsection{Method \code{integrateFractionRemoved()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$integrateFractionRemoved(
  lower = self$tauMin,
  upper = self$tauMax
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fractionRemaining"></a>}}
\if{latex}{\out{\hypertarget{method-fractionRemaining}{}}}
\subsection{Method \code{fractionRemaining()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$fractionRemaining(tau)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-integrateFractionRemaining"></a>}}
\if{latex}{\out{\hypertarget{method-integrateFractionRemaining}{}}}
\subsection{Method \code{integrateFractionRemaining()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$integrateFractionRemaining(
  lower = self$tauMin,
  upper = self$tauMax
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResTmWtdFracRemovStrg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
