% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\name{StreamCell}
\alias{StreamCell}
\title{Class StreamCell (R6)}
\value{
The ojbect of class \code{StreamCell}
}
\description{
Instantiate a \code{StreamCell} object.  \code{StreamCell}
  inherits from class \code{Cell}.
}
\section{Super class}{
\code{\link[SoilsToStreams:Cell]{SoilsToStreams::Cell}} -> \code{StreamCell}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StreamCell$new()}}
\item \href{#method-clone}{\code{StreamCell$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamCell$new(..., waterVolume, soluteMass, soluteConcentration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{waterVolume}}{is the volume of water in the cell.}

\item{\code{soluteMass}}{is the mass of the solute in the cell.}

\item{\code{soluteConcentration}}{will be calculated from the
\code{waterVolume} and \code{soluteMass}.  Therefore, it is the end user's
responsibility to ensure that the units of volume and mass are consistent
throughout their simulation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamCell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
