% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Cell.R
\docType{data}
\name{StreamCell}
\alias{StreamCell}
\title{Class StreamCell (R6)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StreamCell
}
\arguments{
\item{solute}{concentration is the concentration of the solute in user
specified units.}

\item{discharge}{is the water discharge, Q, in user specified units.}

\item{alpha}{is the exponent of the power law used to represent the shape of
the residence time distribution of the hyporheic zone.}

\item{aquiferVolume}{is the volume of the stream aquifer/hyporheic zone.}

\item{porosity}{is the porosity of the streambed and hyporheic zone.}

\item{channelWidth}{is the width of the stream channel surface.}

\item{channelLength}{is the length of the stream channel surface for the
given reach represented by the \code{StreamCell} specified herein.}

\item{channelArea}{is the area of the stream channel surface for the given
\code{StreamCell}.}

\item{channelDepth}{is the height of the water surface above the streambed.}

\item{channelVelocity}{is the mean velocity of the water in the stream
channel.}

\item{channelResidenceTime}{is the mean residence time of the reach
represented by the \code{StreamCell} specified herein.}

\item{qStorage}{is the total hyporheic exchange rate, i.e., the rate at which
water is downwelling into and - since this model assumes steady state -
upwelling up from the hyporheic zone.}

\item{hydraulicLoad}{is the hydraulic load for the reach represented by the
\code{StreamCell} specified herein.}

\item{tauMin}{is the minimum residence time for the hyporheic zone.}

\item{tauMax}{is the maximum residence time for the hyporheic zone.}
}
\value{
The object of class \code{Cell}.
}
\description{
Instantiate a \code{StreamCell} object. Class \code{StreamCell}
  inherits from class \code{Cell}. The following parameters will be
  calculated on the fly from the remaining parameters if they are not
  specified directly by the user: \code{channelArea, channelVelocity,
  channelResidenceTime, qStorage, hydraulicLoad}.
}
\keyword{datasets}
