% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_WQModel.R
\name{WQModel}
\alias{WQModel}
\title{Class WQModel (R6)}
\value{
The ojbect of class \code{WQModel}.
}
\description{
Define the WQ model and the network topology of cells and
  boundaries
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WQModel$new()}}
\item \href{#method-clone}{\code{WQModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WQModel$new(boundsTable, cellsTable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boundsTable}}{At this point, \code{booundsTable} needs to be a table
specifying the currency, type of boundary (reaction or transport), the
boundaryIdx (name as character), the upstreamCellIdx, the
downstreamCellIdx, and the calculateOrder.}

\item{\code{cellsTable}}{must contain the processDomain, cellIdx (as character) and
all other neccessary values to have a cell representing a process domain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WQModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
